CR-869 RHUI Unlinked UAC Incomplete - Removed linking successful page
