+++ app/common_handlers.py
@@ -405,8 +405,15 @@ class CommonConfirmAddress(CommonCommon):
 
                     self.validate_case(uprn_return)
 
-                    raise HTTPFound(
-                        request.app.router['StartAddressHasBeenLinked:get'].url_for(display_region=display_region))
+                    if session['case']['region'][0] == 'N':
+                        raise HTTPFound(
+                            request.app.router['StartNILanguageOptions:get'].url_for())
+                    else:
+                        attributes['language'] = locale
+                        attributes['display_region'] = display_region
+                        await self.call_questionnaire(request, session['case'],
+                                                      attributes, request.app,
+                                                      session.get('adlocation'))
 
                 except ClientResponseError as ex:
                     if ex.status == 404:
